make it easier to provide your own executor
