+++ CMakeLists.txt
@@ -43,7 +43,6 @@ endif()
 
 add_library( appbase
              application_base.cpp
-             application.cpp
              ${HEADERS}
            )
