+++ include/appbase/application.hpp
@@ -1,73 +1,24 @@
+// ---------------------------------------------------------------------------------------------
+// to use a different executor:
+//
+//   1. create a copy of this file in your own application
+//
+//   2. create your own version of appbase/default_executor.hpp
+//           (it needs to define a type `appbase_executor`)
+//      and include this file instead of <appbase/default_executor.hpp>
+//
+//   3. include your own `application.hpp` in your project
+//
+// ---------------------------------------------------------------------------------------------
+
+
+
 #pragma once
 
 #include <appbase/application_base.hpp>
 
+#include <appbase/default_executor.hpp>
 
-namespace appbase {
+#include <appbase/application_instance.hpp>
 
-   using executor = default_executor;
-
-   class application : private executor, public application_base {
-   public:
-      static application&  instance();
-
-      static void reset_app_singleton() { app_instance.reset(); }
-      
-      static bool null_app_singleton()  { return !app_instance; }
-
-      template <typename Func>
-      auto post( int priority, Func&& func ) {
-         return application_base::post(*static_cast<executor*>(this), priority, std::forward<Func>(func));
-      }
-
-      void exec() {
-         application_base::exec(*static_cast<executor*>(this));
-      }
-
-      void startup() {
-         application_base::startup(get_io_service());
-      }
-
-      application() {
-         set_stop_executor_cb([&]() { get_io_service().stop(); });
-         set_post_cb([&](int prio, std::function<void()> cb) { this->post(prio, std::move(cb)); });
-      }
-
-  private:
-      inline static std::unique_ptr<application> app_instance;      
-   };
-}
-
-#include <appbase/plugin.hpp>
-
-namespace appbase {
-
-   template<typename Data, typename DispatchPolicy>
-   void channel<Data,DispatchPolicy>::publish(int priority, const Data& data) {
-      if (has_subscribers()) {
-         // this will copy data into the lambda
-         app().post( priority, [this, data]() {
-            _signal(data);
-         });
-      }
-   }
-
-    class scoped_app {
-   public:
-      explicit scoped_app()  { assert(application::null_app_singleton()); app_ = &app(); }
-      ~scoped_app() { application::reset_app_singleton(); } // destroy app instance so next instance gets a clean one
-
-      scoped_app(const scoped_app&) = delete;
-      scoped_app& operator=(const scoped_app&) = delete;
-
-      // access methods
-      application*       operator->()       { return app_; }
-      const application* operator->() const { return app_; }
-
-   private:
-      application* app_;
-   };
-   
-
-}
